+++ app/Http/Controllers/RepositoryController.php
@@ -27,6 +27,10 @@ class RepositoryController extends Controller
 
     public function create($project_id)
     {
+        if(!auth()->user()->can('add_edit_repositories')) {
+            abort(403);
+        }
+
         $project = Project::findOrFail($project_id);
         return view('repository.create_page')
             ->with('project', $project);
@@ -46,6 +50,10 @@ class RepositoryController extends Controller
 
     public function edit($project_id, $repository_id)
     {
+        if(!auth()->user()->can('add_edit_repositories')) {
+            abort(403);
+        }
+
         $project = Project::findOrFail($project_id);
         $repository = Repository::findOrFail($repository_id);
 
@@ -60,6 +68,10 @@ class RepositoryController extends Controller
 
     public function store(Request $request)
     {
+        if(!auth()->user()->can('add_edit_repositories')) {
+            abort(403);
+        }
+
         $request->validate([
             'title' => 'required',
             'project_id' => 'required',
@@ -79,6 +91,10 @@ class RepositoryController extends Controller
 
     public function update(Request $request)
     {
+        if(!auth()->user()->can('add_edit_repositories')) {
+            abort(403);
+        }
+
         $repository = Repository::findOrFail($request->id);
 
         $repository->title = $request->title;
@@ -93,6 +109,10 @@ class RepositoryController extends Controller
 
     public function destroy(Request $request)
     {
+        if(!auth()->user()->can('delete_repositories')) {
+            abort(403);
+        }
+
         $repository = Repository::findOrFail($request->id);
         $repository->delete();
         return redirect()->route('repository_list_page', $request->project_id);
